requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.24
djangorestframework==3.11.2
djangorestframework-simplejwt==4.0.0
idna==2.10
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
